+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 0.7.0
+* + : RSS Feeds detection
 * + : Translation fix
 
 # 0.6.8